agregar registros

+++ src/cocolab/Models/Registro.cs
@@ -8,7 +8,7 @@ namespace cocolab.Models
     public class Registro
     {
         public virtual int IdRegistro { get; set; }
-        public virtual int NoControlAlumno { get; set; }
+        public virtual long NoControlAlumno { get; set; }
         public virtual DateTime FechaSolicitud { get; set; }
         public virtual DateTime? FechaFinalizacion { get; set; }
         public virtual string Uso { get; set; }
